+++ spec/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/IndexCreatorSpec.php
@@ -15,6 +15,8 @@ use Pim\Bundle\CatalogBundle\Entity\Locale;
 use Pim\Bundle\CatalogBundle\Entity\Repository\CurrencyRepository;
 use Pim\Bundle\CatalogBundle\Entity\Repository\LocaleRepository;
 use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
+use Psr\Log\LoggerInterface;
+use Prophecy\Argument;
 
 /**
  * @require Doctrine\ODM\MongoDB\DocumentManager
@@ -26,7 +28,8 @@ class IndexCreatorSpec extends ObjectBehavior
         ManagerRegistry $managerRegistry,
         DocumentManager $documentManager,
         NamingUtility $namingUtility,
-        Collection $collection
+        Collection $collection,
+        LoggerInterface $logger
     ) {
         $managerRegistry->getManagerForClass('Product')->willReturn($documentManager);
         $documentManager->getDocumentCollection('Product')->willReturn($collection);
@@ -34,7 +37,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $this->beConstructedWith(
             $managerRegistry,
             $namingUtility,
-            'Product'
+            'Product',
+            $logger
         );
     }
 
@@ -67,6 +71,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $namingUtility->getScopableAttributes()->willReturn([$title]);
         $namingUtility->getAttributeNormFields($title)->willReturn(['normalizedData.title-ecommerce', 'normalizedData.title-mobile']);
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -107,6 +113,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $namingUtility->getLocalizableAttributes()->willReturn([$description]);
         $namingUtility->getAttributeNormFields($description)->willReturn(['normalizedData.description-en_US', 'normalizedData.description-de_DE']);
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -136,13 +144,14 @@ class IndexCreatorSpec extends ObjectBehavior
         $price->isUseableAsGridFilter()->willReturn(true);
         $price->getAttributeType()->willReturn('pim_catalog_price_collection');
 
-
         $namingUtility->getPricesAttributes()->willReturn([$price]);
         $namingUtility->getCurrencyCodes()->willReturn(['EUR', 'USD']);
         $namingUtility->appendSuffixes(['normalizedData.price', 'normalizedData.price'], ['EUR', 'USD'], '.')->willReturn(['normalizedData.price.EUR', 'normalizedData.price.USD']);
         $namingUtility->appendSuffixes(['normalizedData.price', 'normalizedData.price'], ['data'], '.')->willReturn(['normalizedData.price.EUR.data', 'normalizedData.price.USD.data']);
         $namingUtility->getAttributeNormFields($price)->willReturn(['normalizedData.price', 'normalizedData.price']);
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -165,6 +174,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $name->isUseableAsGridFilter()->willReturn(true);
         $name->getAttributeType()->willReturn('pim_catalog_text');
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -190,6 +201,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $ean->isUseableAsGridFilter()->willReturn(false);
         $ean->getAttributeType()->willReturn('pim_catalog_text');
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -214,6 +227,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $sku->isUseableAsGridFilter()->willReturn(false);
         $sku->isScopable()->willReturn(false);
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -244,6 +259,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $price->isUseableAsGridFilter()->willReturn(true);
         $price->getAttributeType()->willReturn('pim_catalog_price_collection');
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -267,6 +284,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $color->isUseableAsGridFilter()->willReturn(true);
         $color->getAttributeType()->willReturn('pim_catalog_simpleselect');
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -293,6 +312,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $title->isUseableAsGridFilter()->willReturn(true);
         $title->getAttributeType()->willReturn('pim_catalog_simpleselect');
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -317,6 +338,8 @@ class IndexCreatorSpec extends ObjectBehavior
         $description->isUseableAsGridFilter()->willReturn(true);
         $description->getAttributeType()->willReturn('pim_catalog_simpleselect');
 
+        $indexes = array_fill(0, 10, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
         $options =  [
             'background' => true,
             'w'          => 0
@@ -342,6 +365,33 @@ class IndexCreatorSpec extends ObjectBehavior
         $description->getAttributeType()->willReturn('pim_catalog_simpleselect');
 
         $namingUtility->getAttributeNormFields($description)->willReturn(['normalizedData.description-en_US-ecommerce', 'normalizedData.description-de_DE-ecommerce', 'normalizedData.description-en_US-mobile']);
+
+        $this->ensureIndexesFromAttribute($description);
+    }
+
+
+    function it_logs_error_when_the_maximum_number_of_indexes_is_reached(
+        $collection,
+        AbstractAttribute $description,
+        $namingUtility,
+        $logger
+    ) {
+        $description->getCode()->willReturn('description');
+        $description->getBackendType()->willReturn('varchar');
+        $description->isLocalizable()->willReturn(true);
+        $description->isScopable()->willReturn(false);
+        $description->isUseableAsGridFilter()->willReturn(true);
+        $description->getAttributeType()->willReturn('pim_catalog_textarea');
+
+        $namingUtility
+            ->getAttributeNormFields($description)
+            ->willReturn(['normalizedData.description-en_US', 'normalizedData.description-de_DE']);
+
+        $indexes = array_fill(0, 64, 'fake_index');
+        $collection->getIndexInfo()->willReturn($indexes);
+
+        $logger->error(Argument::any())->shouldBeCalled();
+        $collection->ensureIndex(Argument::any(), Argument::any())->shouldNotBeCalled();
 
         $this->ensureIndexesFromAttribute($description);
     }